Feat IMPORTANT: Turbo streams queue on client end
Previously, Ruby would sleep the thread of streams to prevent overlap.
I decided I did not like this method. Turbo does offer an event listener
for pre-render of a stream. However, unlike the standard Turbo render,
this event does not seem to allow you to resume render.
Due to this, I wrote a script that intercepts streams and will pause any
overlapping streams and queue them to be manually inserted into the DOM